+++ NEWS.md
@@ -23,6 +23,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 * Rename `cjo` to `td` in functions name (`assign_cjo`, `export_cjo`, `retrieve_cjo`, `import_cjo`, ) and argument (`cjo` -> `td`)
 * The `assign_XXX` and `retrieve_XXX` functions are using `jws` instead of the `ws_path`.
 * `select_regs` takes a new argument `context` to define the different regressors sets to use.
+* `export_XXX` and `import_XXX` functions don't support anymore `ws_name` argument.
 
 ### Fixed