+++ NEWS.md
@@ -25,10 +25,11 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 * `select_regs` takes a new argument `context` to define the different regressors sets to use.
 * `export_XXX` and `import_XXX` functions don't support anymore `ws_name` argument.
 * `export_XXX` functions are using `td` or `outliers` to replace `x`.
+* `get_LY_info` works now with a X13 model and not a summary_x13 model
 
 ### Fixed
 
-* Bug in select regressors solved
+* Bug in select regressors solved (when no TD is significative)
 
 
 ## [1.0.0] - 2025-08-29